src/msty-millionaire-backend/requirements.txt
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
eventlet==0.40.0
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
redis==6.2.0
requests==2.32.4
simple-websocket==1.1.0
SQLAlchemy==2.0.41
stripe==12.2.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
